[Fix] 해님달님 intro 보이스오버 기능 안 되는 문제 해결
