+++ Codingdong-iOS/Codingdong-iOS/Presentation/SM_Intro/ViewController/SunAndMoonIntroViewController.swift
@@ -27,8 +27,80 @@ final class SunAndMoonIntroViewController: UIViewController, ConfigUI {
         return label
     }()
     
-    private let labelComponents = SunAndMoonIntroView()
+    // 지우기
+    //private let labelComponents = SunAndMoonIntroView()
+    private let introLabel: UILabel = {
+        let label = UILabel()
+        label.text = "해님달님에서 아래의 코딩 개념들을 배워 보아요!"
+        label.isAccessibilityElement = true
+        label.font = FontManager.body()
+        label.textColor = .gs10
+        label.numberOfLines = 0
+        label.lineBreakMode = .byCharWrapping
+        return label
+    }()
     
+    private let firstConceptLabel: UILabel = {
+        let label = UILabel()
+        label.text = "개념 1. 조건문"
+        label.font = FontManager.title3()
+        label.textColor = .gs10
+        label.numberOfLines = 0
+        label.lineBreakMode = .byCharWrapping
+        return label
+    }()
+    
+    private let firstDescriptionLabel: UILabel = {
+        let label = UILabel()
+        label.text = "호랑이랑 마주친 엄마"
+        label.font = FontManager.caption2()
+        label.textColor = .gs30
+        label.numberOfLines = 0
+        label.lineBreakMode = .byCharWrapping
+        return label
+    }()
+    
+    private let secondConceptLabel: UILabel = {
+        let label = UILabel()
+        label.text = "개념 2. 연산자"
+        label.font = FontManager.title3()
+        label.textColor = .gs10
+        label.numberOfLines = 0
+        label.lineBreakMode = .byCharWrapping
+        return label
+    }()
+    
+    private let secondDescriptionLabel: UILabel = {
+        let label = UILabel()
+        label.text = "남매의 집에 도착한 호랑이"
+        label.font = FontManager.caption2()
+        label.textColor = .gs30
+        label.numberOfLines = 0
+        label.lineBreakMode = .byCharWrapping
+        return label
+    }()
+    
+    private let thirdConceptLabel: UILabel = {
+        let label = UILabel()
+        label.text = "개념 3. 반복문"
+        label.font = FontManager.title3()
+        label.textColor = .gs10
+        label.numberOfLines = 0
+        label.lineBreakMode = .byCharWrapping
+        return label
+    }()
+    
+    private let thirdDescriptionLabel: UILabel = {
+        let label = UILabel()
+        label.text = "동아줄을 잡은 남매와 호랑이"
+        label.font = FontManager.caption2()
+        label.textColor = .gs30
+        label.numberOfLines = 0
+        label.lineBreakMode = .byCharWrapping
+        return label
+    }()
+    
+
     private let nextButton = CommonButton()
     
     private lazy var nextButtonViewModel = CommonbuttonModel(title: "시작하기", font: FontManager.textbutton(), titleColor: .primary1, backgroundColor: .gs10, height: 72) {
@@ -65,16 +137,48 @@ final class SunAndMoonIntroViewController: UIViewController, ConfigUI {
     }
     
     func addComponents() {
-        [labelComponents, nextButton].forEach {
+        [introLabel, firstConceptLabel, firstDescriptionLabel, secondConceptLabel, secondDescriptionLabel, thirdConceptLabel, thirdDescriptionLabel, nextButton].forEach {
             view.addSubview($0)
         }
     }
     
     func setConstraints() {
-        labelComponents.snp.makeConstraints {
-            $0.left.equalToSuperview()
-            $0.right.equalToSuperview()
-            $0.top.equalTo(naviLine).offset(basicPadding)
+        introLabel.snp.makeConstraints {
+            $0.top.equalTo(naviLine.snp.bottom).offset(16)
+            $0.left.right.equalToSuperview().inset(16)
+        }
+        
+        firstConceptLabel.snp.makeConstraints {
+            $0.top.equalTo(introLabel.snp.bottom).offset(32)
+            $0.left.right.equalToSuperview().inset(16)
+        }
+        
+        firstDescriptionLabel.snp.makeConstraints {
+            $0.top.equalTo(firstConceptLabel.snp.bottom).offset(8)
+            $0.left.right.equalToSuperview().inset(16)
+        }
+        
+        secondConceptLabel.snp.makeConstraints {
+            $0.top.equalTo(firstDescriptionLabel.snp.bottom).offset(20)
+            $0.left.right.equalToSuperview().inset(16)
+        }
+        
+        
+        secondDescriptionLabel.snp.makeConstraints {
+            $0.top.equalTo(secondConceptLabel.snp.bottom).offset(8)
+            $0.left.right.equalToSuperview().inset(16)
+        }
+        
+        
+        thirdConceptLabel.snp.makeConstraints {
+            $0.top.equalTo(secondDescriptionLabel.snp.bottom).offset(20)
+            $0.left.right.equalToSuperview().inset(16)
+        }
+        
+        
+        thirdDescriptionLabel.snp.makeConstraints {
+            $0.top.equalTo(thirdConceptLabel.snp.bottom).offset(8)
+            $0.left.right.equalToSuperview().inset(16)
         }
         
         nextButton.snp.makeConstraints {
@@ -86,7 +190,7 @@ final class SunAndMoonIntroViewController: UIViewController, ConfigUI {
     
     func setupAccessibility() {
         self.navigationItem.leftBarButtonItem?.accessibilityLabel = "이전 화면으로 이동"
-        view.accessibilityElements = [labelComponents, nextButton]
+        view.accessibilityElements = [introLabel, firstConceptLabel, firstDescriptionLabel, secondConceptLabel, secondDescriptionLabel, thirdConceptLabel, thirdDescriptionLabel, nextButton]
         
     }
